+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE-GLOBAL-WITH-NEXT-VERSION
+- PT-11910 - Add migration of cross selling
+
 # 1.6.0
 - PT-11692 - Add functionality to finish migration and remove unneeded data
 - PT-11864 - Improve media file processing